python/requirements.txt
reduct-py>=1.15